requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
nltk==3.8.1
pandas==2.0.3
numpy>=1.24.3,<2.0.0
joblib==1.3.2
openai==1.3.0
python-dotenv==1.0.0
spellchecker==0.4
requests==2.31.0
gunicorn==21.2.0
transformers==4.35.0
torch==2.1.0
